+++ packages/vtable/src/index.ts
@@ -1,4 +1,6 @@
 /* eslint-disable sort-imports */
+import { graphicUtil } from '@visactor/vrender';
+
 import * as TYPES from './ts-types';
 import * as core from './core';
 import * as data from './data';
@@ -32,6 +34,9 @@ import { PivotTable } from './PivotTable';
 import { PivotChart } from './PivotChart';
 import type { MousePointerCellEvent } from './ts-types/events';
 import * as CustomLayout from './render/layout';
+
+import { updateCell } from './scenegraph/group-creater/cell-helper';
+import { renderChart } from './scenegraph/graphic/contributions/chart-render-helper';
 
 // import { container, loadCanvasPicker } from '@visactor/vrender';
 // loadCanvasPicker(container);
@@ -83,7 +88,10 @@ export {
    * 暂不推荐使用
    */
   DataStatistics,
-  CustomLayout
+  CustomLayout,
+  updateCell,
+  renderChart,
+  graphicUtil
 };
 
 /** @private */