+++ packages/vtable/src/scenegraph/layout/update-width.ts
@@ -29,15 +29,23 @@ import { getHierarchyOffset } from '../utils/get-hierarchy-offset';
  * @param {number} detaX
  * @return {*}
  */
-export function updateColWidth(scene: Scenegraph, col: number, detaX: number) {
-  scene.table._setColWidth(col, scene.table.getColWidth(col) + detaX, true);
+export function updateColWidth(scene: Scenegraph, col: number, detaX: number, skipTableWidthMap?: boolean) {
+  if (!skipTableWidthMap) {
+    scene.table._setColWidth(col, scene.table.getColWidth(col) + detaX, true);
+  }
 
   const autoRowHeight = scene.table.heightMode === 'autoHeight';
   // deal with corner header or column header
   const colOrCornerHeaderColumn = scene.getColGroup(col, true) as Group;
-  if (colOrCornerHeaderColumn) {
+  const rightTopColumn = scene.getColGroupInRightTopCorner(col);
+  if (colOrCornerHeaderColumn && !rightTopColumn) {
     updateColunmWidth(colOrCornerHeaderColumn, detaX, autoRowHeight, 'col-corner', scene);
   }
+  // deal with right bottom frozen cells
+  if (rightTopColumn) {
+    updateColunmWidth(rightTopColumn, detaX, autoRowHeight, 'right-top', scene);
+  }
+
   // deal with row header or body or right frozen cells
   const rowHeaderOrBodyColumn = scene.getColGroup(col) as Group;
   if (rowHeaderOrBodyColumn) {
@@ -53,11 +61,6 @@ export function updateColWidth(scene: Scenegraph, col: number, detaX: number) {
   const bottomColumn = scene.getColGroupInBottom(col);
   if (bottomColumn) {
     updateColunmWidth(bottomColumn, detaX, autoRowHeight, 'bottom', scene);
-  }
-  // deal with right bottom frozen cells
-  const rightTopColumn = scene.getColGroupInRightTopCorner(col);
-  if (rightTopColumn) {
-    updateColunmWidth(rightTopColumn, detaX, autoRowHeight, 'right-top', scene);
   }
   // deal with right bottom frozen cells
   const rightBottomColumn = scene.getColGroupInRightBottomCorner(col);
@@ -220,9 +223,10 @@ function updateCellWidth(
   // autoColWidth: boolean,
   autoRowHeight: boolean
 ): boolean {
-  if (cell.attribute.width === distWidth) {
+  if (cell.attribute.width === distWidth && !cell.needUpdateWidth) {
     return false;
   }
+  cell.needUpdateWidth = false;
 
   cell.setAttribute('width', distWidth);
   // const mergeInfo = getCellMergeInfo(scene.table, col, row);
@@ -438,7 +442,7 @@ function updateMergeCellContentWidth(
         // const { width: contentWidth } = cellGroup.attribute;
         singleCellGroup.contentWidth = distWidth;
 
-        resizeCellGroup(
+        const { heightChange } = resizeCellGroup(
           singleCellGroup,
           rangeWidth,
           rangeHeight,
@@ -454,6 +458,11 @@ function updateMergeCellContentWidth(
           },
           table
         );
+
+        if (heightChange) {
+          singleCellGroup.needUpdateHeight = true;
+        }
+
         isHeightChange = isHeightChange || changed;
       }
     }