+++ packages/vtable/CHANGELOG.md
@@ -1,6 +1,112 @@
 # Change Log - @visactor/vtable
 
-This log was last generated on Thu, 21 Dec 2023 11:54:58 GMT and should not be manually modified.
+This log was last generated on Fri, 05 Jan 2024 10:34:08 GMT and should not be manually modified.
+
+## 0.17.8
+Fri, 05 Jan 2024 10:34:08 GMT
+
+### Updates
+
+- fix: selectRange error when near frozencol or frozenrow #854
+
+
+- fix:  frozen shadowline should move position #859
+
+
+- fix: fix chart cell dblclick size update
+- fix: fix bottom frozen row height compute in createGroupForFirstScreen()
+- fix: fix cellGroup merge range
+- fix: fix react custom jsx parse
+
+## 0.17.7
+Thu, 04 Jan 2024 10:03:11 GMT
+
+### Updates
+
+- feat: add cell image table export
+- fix: fix jsx parse error in react-vtable
+
+## 0.17.6
+Wed, 03 Jan 2024 13:58:45 GMT
+
+### Updates
+
+- fix: fix resize line position
+
+## 0.17.5
+Wed, 03 Jan 2024 12:50:32 GMT
+
+### Updates
+
+- feat: support edit header title #819
+
+
+- fix: setRecords process scrollTop update scenegraph #831
+
+
+- refactor: list table bottom row can not use bottomFrozenStyle #836
+
+
+- feat: add api getCellHeaderTreeNodes for pivotTable #839
+
+
+- refactor: add onVChartEvent for BaseTable #843
+
+
+- fix: add group clip in body
+
+## 0.17.4
+Tue, 02 Jan 2024 09:23:54 GMT
+
+### Updates
+
+- fix: fix frozen chart cell active problem 
+
+## 0.17.3
+Fri, 29 Dec 2023 12:31:07 GMT
+
+### Updates
+
+- fix: setRecords lose hover state  #783
+
+
+- feat: add body index convert with table index #789
+
+
+- fix:  transpose list demo when records has 10000 performance problem #790
+
+
+- fix: setRecords recomputeColWidth problems #796
+
+
+- fix: set disableSelect drag interaction occor error #799
+
+
+- feat: mergeCell support custom compare function #804
+
+
+- fix: tooltip style not work #805
+
+
+- fix: pivot table pagination.perPageCount modify #807
+
+
+- fix: [Bug] adaptive mode compute problem when has frozencol and rightFrozenCol #820
+
+
+- fix: fix axis render update problem
+- feat: add column resize label theme
+- fix: fix select update when change frozen
+- fix: pivot table use icon bug
+
+
+- fix: fix sort icon update
+- refactor: update vrender event verison use scrollDrag
+
+
+- chore: update vrender version #785
+
+
 
 ## 0.17.2
 Thu, 21 Dec 2023 11:54:58 GMT